Extract readiness operability helper
